댓글 기능 인증 추가

+++ src/main/java/com/back/domain/post/postComment/controller/ApiV1PostCommentController.java
@@ -7,8 +7,13 @@ import com.back.domain.post.post.service.PostService;
 import com.back.domain.post.postComment.dto.*;
 import com.back.domain.post.postComment.entity.PostComment;
 import com.back.domain.post.postComment.service.PostCommentService;
+import com.back.global.exception.ServiceException;
 import com.back.global.rsData.RsData;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
@@ -19,6 +24,7 @@ import java.util.stream.Collectors;
 @RestController // @Controller + @ResponseBody
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/posts/{postId}/comments")
+@Tag(name = "ApiV1PostCommentController", description = "API 댓글 컨트롤러")
 public class ApiV1PostCommentController {
     private final PostService postService;
     private final PostCommentService postCommentService;
@@ -26,14 +32,21 @@ public class ApiV1PostCommentController {
 
     @GetMapping("")
     @Transactional(readOnly = true)
-    public List<PostCommentDto> getComments(@PathVariable Long postId) {
+    @Operation(summary = "다건 조회")
+    public List<PostCommentDto> getComments(
+            @PathVariable Long postId
+    ) {
         Post post = postService.findById(postId);
         return post.getPostComments().stream().map(PostCommentDto::new).collect(Collectors.toList());
     }
 
     @GetMapping("/{id}")
     @Transactional(readOnly = true)
-    public PostCommentDto getComment(@PathVariable Long postId, @PathVariable Long id) {
+    @Operation(summary = "단건 조회")
+    public PostCommentDto getComment(
+            @PathVariable Long postId,
+            @PathVariable Long id
+    ) {
         Post post = postService.findById(postId);
 
         return new PostCommentDto(postCommentService.getCommentById(post, id));
@@ -41,9 +54,20 @@ public class ApiV1PostCommentController {
 
     @DeleteMapping("/{id}")
     @Transactional
-    public RsData<Void> delete(@PathVariable Long postId, @PathVariable Long id) {
+    @Operation(summary = "삭제")
+    public RsData<Void> delete(
+            @PathVariable Long postId,
+            @PathVariable Long id,
+            @NotBlank @Size(min = 2, max = 50) @RequestHeader("Authorization") String authorization
+    ) {
+        String apiKey = authorization.replace("Bearer ", "");
+        Member author = memberService.findByApiKey(apiKey).get();
         Post post = postService.findById(postId);
         PostComment postComment = postCommentService.getCommentById(post, id);
+        if (!author.equals(postComment.getAuthor())) {
+            throw new ServiceException("403-1", "작성자만 댓글을 삭제할 수 있습니다.");
+        }
+
         postCommentService.delete(post, postComment);
 
         return new RsData<>("200-1", "%d번 댓글이 삭제되었습니다.".formatted(id));
@@ -51,20 +75,47 @@ public class ApiV1PostCommentController {
 
     @PostMapping("")
     @Transactional
-    public RsData<PostCommentWriteResBody> write(@PathVariable Long postId, @Valid @RequestBody PostCommentWriteReqBody reqBody) {
-        Member member = memberService.findByUsername("user1").get();
+    @Operation(summary = "작성")
+    public RsData<PostCommentDto> write(
+            @PathVariable Long postId,
+            @Valid @RequestBody PostCommentWriteReqBody reqBody,
+            @NotBlank @Size(min = 2, max = 50) @RequestHeader("Authorization") String authorization
+    ) {
+        String apiKey = authorization.replace("Bearer ", "");
+        Member author = memberService.findByApiKey(apiKey).get();
         Post post = postService.findById(postId);
-        PostComment postComment = postCommentService.create(member, post, reqBody.content());
-        return new RsData<>("201-1", "%d번 댓글이 생성되었습니다.".formatted(postComment.getId()), new PostCommentWriteResBody(postCommentService.countPostCommentsByPost(post), new PostCommentDto(postComment)));
+        System.out.println(post);
+        PostComment postComment = postCommentService.create(author, post, reqBody.content());
+        return new RsData<>(
+                "201-1",
+                "%d번 댓글이 생성되었습니다.".formatted(postComment.getId()),
+                new PostCommentDto(postComment)
+        );
     }
 
     @PutMapping("/{id}")
     @Transactional
-    public RsData<PostCommentUpdateResDto> update(@PathVariable Long postId, @PathVariable Long id, @Valid @RequestBody PostCommentUpdateReqDto reqBody) {
+    @Operation(summary = "수정")
+    public RsData<PostCommentDto> update(
+            @PathVariable Long postId,
+            @PathVariable Long id,
+            @Valid @RequestBody PostCommentUpdateReqDto reqBody,
+            @NotBlank @Size(min = 2, max = 50) @RequestHeader("Authorization") String authorization
+    ) {
+        String apiKey = authorization.replace("Bearer ", "");
+        Member author = memberService.findByApiKey(apiKey).get();
         Post post = postService.findById(postId);
         PostComment postComment = postCommentService.getCommentById(post, id);
+        if (!author.equals(postComment.getAuthor())) {
+            throw new ServiceException("403-1", "작성자만 댓글을 수정할 수 있습니다.");
+        }
+
         postCommentService.update(post, postComment, reqBody.content());
 
-        return new RsData<>("200-1", "%d번 댓글이 수정되었습니다.".formatted(post.getId()), new PostCommentUpdateResDto(new PostCommentDto(postComment)));
+        return new RsData<>(
+                "200-1",
+                "%d번 댓글이 수정되었습니다.".formatted(post.getId()),
+                new PostCommentDto(postComment)
+        );
     }
 }
